fix(themes.astro): change script to not be a module
Script modules have deferred execution and run after all non-module scripts on a page. This is
undesirable for a script we want to immediately run, especially if we want to run before render.
After this fix is made the playwright test in the previous commit starts working (the theme script
runs immediately)